#11 追加テーブルへのデータ保存

+++ www/model/cart.php
@@ -123,20 +123,65 @@ function purchase_carts($db, $carts){
     // FALSEを返す
     return false;
   }
+
   // カート内の商品が購入できる場合
+  // トランザクション開始
+  $db->beginTransaction();
+  // 購入履歴データを登録
+  if (insert_order($db, $carts[0]['user_id']) === false){
+    // FALSEを返す
+    return false;
+  }  
+  // order_idの値を取得
+  $order_id = $db->lastInsertId('order_id');
+  // 繰り返し処理で購入の処理
   foreach($carts as $cart){
-    // 在庫数の更新に失敗した場合
-    if(update_item_stock(
-        $db, 
-        $cart['item_id'], 
-        $cart['stock'] - $cart['amount']
-      ) === false){
+    // 在庫数の更新及び購入明細データの登録に失敗した場合
+    if(update_item_stock($db, $cart['item_id'], $cart['stock'] - $cart['amount']) === false
+      || insert_order_detail($db, $order_id, $cart['item_id'], $cart['price'], $cart['amount']) === false){
+
       // エラーメッセージを定義
-      set_error($cart['name'] . 'の購入に失敗しました。');
+      // set_error($cart['name'] . 'の購入に失敗しました。');
+
+      // ロールバック処理(取り消し)
+      $db->rollback();
+      // FALSEを返す
+      return false;
     }
   }
+  // コミット処理
+  $db->commit();
   // 購入したユーザのカート内の全商品を削除
   delete_user_carts($db, $carts[0]['user_id']);
+}
+
+// 購入履歴データを登録
+function insert_order($db, $user_id){
+  $sql = "
+    INSERT INTO
+      orders(
+        user_id
+      )
+    VALUES(?);
+  ";
+  // sqlを実行して成功した場合に true を、失敗した場合に false を返す
+  return execute_query($db, $sql, [$user_id]);
+}
+
+// 購入明細データを登録
+function insert_order_detail($db, $order_id, $item_id, $order_price, $order_amount){
+  $sql = "
+    INSERT INTO
+      order_details(
+        order_id,
+        item_id,
+        order_price,
+        order_amount
+      )
+    VALUES(?, ?, ?, ?);
+  ";
+  // sqlを実行して成功した場合に true を、失敗した場合に false を返す
+  return execute_query($db, $sql, [$order_id, $item_id, $order_price, $order_amount]);
 }
 
 // 特定のユーザのカート内の全商品を削除
